requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
pandas==2.1.4
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
